tools/requirements.txt
cryptography==46.0.3
python-jose[cryptography]==3.5.0
tqdm==4.67.1
